# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOUL V1 API",
    "version": "1.0.0",
    "description": "Machine-readable companion to docs/FLUTTER_API_HANDOFF.md."
  },
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Local development; replace the origin per environment"
    }
  ],
  "paths": {
...
    "/catalogs/profile": {
      "get": {
        "operationId": "api.v1.catalogs.profile",
        "summary": "Localized interests, traits and help categories",
        "tags": [
          "Catalogs"
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/onboarding/religion-profile": {
      "get": {
        "operationId": "api.v1.onboarding.religion-profile.show",
        "summary": "Resume saved selection",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "put": {
        "operationId": "api.v1.onboarding.religion-profile.store",
        "summary": "Save complete hierarchy path",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/onboarding/profile": {
      "get": {
        "operationId": "api.v1.onboarding.profile.show",
        "summary": "Resume draft profile",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "put": {
        "operationId": "api.v1.onboarding.profile.update",
        "summary": "Partially update draft",
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "first_name": "Ayesha",
                "marital_status": "never_married",
                "profession_status": "employed",
                "smoking": "no",
                "alcohol": "no",
                "current_children": "no",
                "future_children": "want_children",
                "intentions": [
                  "marriage"
                ],
                "spoken_language_ids": [
                  1
                ],
                "interests": [
                  "Reading"
                ],
                "personality_traits": [
                  "Kind"
                ],
                "prefer_not_to_say_fields": []
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/profiles/{profile}": {
      "get": {
        "operationId": "api.v1.profiles.show",
        "summary": "Safe full public profile with prominent marital status",
        "tags": [
          "Profiles"
        ],
        "parameters": [
          {
            "name": "profile",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/private-photos/{photo}/content": {
      "get": {
        "operationId": "api.v1.private-photos.content",
        "summary": "Authorized no-store image bytes",
        "tags": [
          "Private-photos"
        ],
        "parameters": [
          {
            "name": "photo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "description": "Authorized private image bytes",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/webhooks/cloudinary/moderation": {
      "post": {
        "operationId": "api.v1.webhooks.cloudinary.moderation",
        "summary": "Ingest signed Cloudinary moderation notification",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "cloudinarySignature": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum token"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session"
      },
      "cloudinarySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cld-Signature"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful SOUL response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
      "Error": {
        "description": "Standard SOUL error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}